リポジトリ層を db.ts に統合
