+++ app/api/routes/login.ts
@@ -4,7 +4,7 @@ import { compare } from "bcrypt"; // bcrypt で検証
 import { z } from "zod";
 import { zValidator } from "@hono/zod-validator";
 import { getEnv } from "../../shared/config.ts";
-import { createSession } from "../repositories/session.ts";
+import { createSession } from "../db.ts";
 
 const app = new Hono();
