+++ app/api/routes/nodeinfo.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
 import { createDB } from "../db.ts";
-import { countAccounts } from "../repositories/account.ts";
+import { countAccounts } from "../db.ts";
 import { getDomain } from "../utils/activitypub.ts";
 import { getEnv } from "../../shared/config.ts";
 // NodeInfo は外部からの参照を想定しているため認証は不要
