+++ app/api/routes/root_inbox.ts
@@ -6,7 +6,7 @@ import {
 } from "../utils/activitypub.ts";
 import { getEnv } from "../../shared/config.ts";
 import { activityHandlers } from "../activity_handlers.ts";
-import { findAccountByUserName } from "../repositories/account.ts";
+import { findAccountByUserName } from "../db.ts";
 import { createDB } from "../db.ts";
 import { addInboxEntry } from "../services/inbox.ts";
