+++ app/api/routes/session.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
 import { getCookie } from "hono/cookie";
-import { deleteSessionById, findSessionById } from "../repositories/session.ts";
+import { deleteSessionById, findSessionById } from "../db.ts";
 import authRequired from "../utils/auth.ts";
 import { getEnv } from "../../shared/config.ts";
