+++ app/api/routes/setup_ui.ts
@@ -2,7 +2,7 @@ import { Hono } from "hono";
 import { load, stringify } from "jsr:@std/dotenv";
 import { ensureFile } from "jsr:@std/fs/ensure-file";
 import { join } from "jsr:@std/path";
-import { createAccount } from "../repositories/account.ts";
+import { createAccount } from "../db.ts";
 import { createDB } from "../db.ts";
 import { getEnv } from "../../shared/config.ts";
 import authRequired from "../utils/auth.ts";
